Add files via upload
Those pascal headers for AVR microcontrollers is not made by me. I found on github a user which has made a simple program in lazarus which convert the atdf file into pascal units which i consider as more completed than mine. 
I  hope to help anyone to made avr programming easier.  The program should be found in this link 
https://github.com/ccrause/importATDF 
and the user which has more project could be found https://github.com/ccrause